+++ src/components/Header/Header.scss
@@ -118,8 +118,18 @@ header {
       }
     }
 
-    .header_login_button {
-      span {
+    .login-button {
+      display: flex;
+      font-size: 18px;
+      font-weight: 500;
+      padding: 5px 18px;
+      border-radius: 12px;
+      border-color: transparent;
+      background-color: transparent;
+
+      &:hover {
+        background-color: $light-grey;
+        cursor: pointer;
       }
     }
   }
