--- src/components/Modal/Login/Login.scss
@@ -1,182 +1,0 @@
-@import '../../../style/common.scss';
-.login-button-modal{
-  display: flex;
-  padding-left: 0;
-  font-size: 18px;
-  font-weight: 500;
-  padding: 5px 18px;
-  border-radius: 12px;
-  border-color: transparent;
-  background-color: transparent;
-  &:hover{
-      background-color: $light-grey;
-      cursor: pointer;
-  }
-}
-.ReactModalPortal {
-  position: relative;
-  z-index: 9999;
-
-  .login-modal-main-overlay{
-    position: fixed;
-    inset: 0;
-    background-color: rgba(0, 0, 0, 0.5);
-
-    .login-modal-main {
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-      width: auto;
-      border: 1px solid rgb(204, 204, 204);
-      background: white;
-      opacity: 1;
-      overflow: auto;
-      border-radius: 3px;
-      outline: none;
-    }
-  }
-}
-.ReactModal__Body--open {
-  overflow: hidden;
-}
-
-
-
-
-.login-modal-total{
-  width: 40vw;
-  .login-modal-total-div{
-    width: 100%;
-    height: 65vh;
-    .login-left{
-      background-color: $light-grey;
-      height: 100%;
-      position: relative;
-      left: 0;
-      width: 35%;
-      float: left;
-      .login-left-tit{
-        position: relative;
-        top: 30%;
-        text-align: center;
-        .login-img{
-          img{
-            width: 80%;
-          }
-        }
-        .login-insa{
-          span{
-            font-weight: 600;
-            font-size: 1.8em;
-          }
-        }
-      }
-    }   
-    .login-right{
-      width: 60%;
-      float: right;
-      padding-right: 10px;
-      .login-right-tit{
-        .login-right-ment{
-          padding-left: 15px;
-          display: inline-block;
-          padding-top:8vh;
-          .login-right-ment-login{
-            
-            span{
-              font-style: normal;
-              font-weight: 600;
-              font-size: 25px;
-            }
-            img{
-              position: relative;
-              top: 3px;
-              padding-left: 6px;
-              width: 30px;
-            }
-          }
-          .login-right-ment-schoollogin{
-            padding: 10px 0px 20px 0px;
-            span{
-              color: $dark-grey;
-              font-size: 14px;
-              font-weight: 500;
-            }
-          }
-        }
-        .login-right-input-idpw{ text-align: center;
-          input{
-            border:0.5px solid $medium-grey;
-            width: 90%;
-            height: 6vh;
-          }
-          ::placeholder{
-            text-indent:10px;
-            color: $medium-grey;
-            font-weight: 300;
-          }        
-          .login-right-ment-email{                                        
-            padding-bottom: 6px;
-          }
-          .login-right-ment-password{
-            padding-bottom: 15px;
-          }  
-        }
-        .login-right-loginbutton-tit{
-          .login-right-loginbutton-login{
-            text-align: center;
-            button{
-              height: 6vh;
-              border: none;
-              background-color: $main-red;
-              color: white;
-              width: 91%;
-              border-radius: 2px;
-              font-size: 16px;
-              font-weight: 500;
-              &:hover{
-                cursor: pointer;
-              }
-            }
-            padding-bottom: 6px;
-          }
-          .login-right-loginbutton-googlelogin{
-            text-align: center;
-            button{
-              height: 6vh;
-              border: 0.5px solid $medium-grey;
-              background-color: white;
-              width: 91%;
-              font-size: 16px;
-              border-radius: 2px;              
-              img{
-              }
-              &:hover{
-                cursor: pointer;
-              }
-            }
-          }
-        }
-      }
-      .login-right-under{
-        padding-top: 10vh;
-        text-align: center;
-          .login-under-que{
-            color: $medium-grey;
-            font-size: 14px;
-          }
-          .login-under-dap{
-            padding-left: 10px;
-            color: $main-red;
-            font-weight: 500;
-            font-size: 14px;
-          }
-          :hover{
-            cursor: pointer;
-          }
-    
-      }
-    }
-  }
-}
