+++ src/components/Modal/Login/LoginModal.scss
@@ -1,0 +1,179 @@
+@import '../../../style/common.scss';
+
+.ReactModalPortal {
+  position: relative;
+  z-index: 9999;
+
+  .login-modal-main-overlay {
+    position: fixed;
+    inset: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+
+    .login-modal {
+      position: absolute;
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%);
+      border: 1px solid rgb(204, 204, 204);
+      background: white;
+      opacity: 1;
+      overflow: auto;
+      border-radius: 3px;
+      outline: none;
+    }
+  }
+}
+
+.ReactModal__Body--open {
+  overflow: hidden;
+}
+
+
+.login-modal {
+  width: 40vw;
+  height: 70vh;
+
+  .login-left {
+    width: 35%;
+    height: 100%;
+    background-color: $light-grey;
+    position: relative;
+    left: 0;
+    float: left;
+
+    .login-left-tit {
+      position: relative;
+      top: 30%;
+      text-align: center;
+
+      .login-img {
+        img {
+          width: 80%;
+        }
+      }
+
+      .login-insa {
+        span {
+          font-weight: 600;
+          font-size: 1.8em;
+        }
+      }
+    }
+  }
+
+  .login-right {
+    display: flex;
+    flex-direction: column;
+    width: 65%;
+    height: 100%;
+    float: right;
+    padding: 100px 30px 30px 50px;
+
+    .login-right-tit {
+      .login-right-ment {
+        display: inline-block;
+
+        .login-right-ment-login {
+          span {
+            font-weight: 600;
+            font-size: 25px;
+          }
+
+          img {
+            position: relative;
+            top: 3px;
+            width: 30px;
+            padding-left: 4px;
+          }
+        }
+
+        .login-right-ment-login-ment {
+          color: $dark-grey;
+          font-size: 14px;
+          font-weight: 400;
+          padding-top: 16px;
+        }
+      }
+
+      .login-right-input-idpw {
+        text-align: center;
+
+        input {
+          border: 0.5px solid $medium-grey;
+          width: 100%;
+          height: 6vh;
+          padding: 0 15px;
+
+          &:focus {
+            border: 1px solid $medium-grey;
+            outline: none;
+          }
+        }
+
+        ::placeholder {
+          color: $medium-grey;
+          font-weight: 300;
+        }
+
+        .login-right-ment-email {
+          padding-bottom: 6px;
+        }
+
+        .login-right-ment-password {
+          padding-bottom: 20px;
+        }
+      }
+
+      .login-right-login-button-area {
+
+        button {
+          width: 100%;
+          height: 6vh;
+          display: flex;
+          justify-content: center;
+          align-items: center;
+          font-size: 16px;
+          font-weight: 500;
+          border-radius: 2px;
+        }
+
+        .login-right-login-button--basic {
+          margin-bottom: 6px;
+
+          button {
+            background-color: $main-red;
+            color: white;
+            margin-right: 4px;
+          }
+        }
+
+        .login-right-login-button--google {
+          button {
+            word-break: keep-all;
+            border: 0.5px solid $medium-grey;
+            background-color: white;
+            border-radius: 2px;
+          }
+        }
+      }
+    }
+
+    .login-right-under {
+      margin-top: auto;
+      text-align: center;
+      color: $medium-grey;
+      font-size: 13px;
+
+      .login-under-dap {
+        word-break: keep-all;
+        color: $main-red;
+        font-weight: 600;
+      }
+
+      :hover {
+        cursor: pointer;
+      }
+
+    }
+  }
+}
