+++ src/index.css
@@ -30,4 +30,5 @@ a {
 
 button {
     cursor: pointer;
+    border: none;
 }
