+++ src/style/common.scss
@@ -5,3 +5,4 @@ $grey: #E6E6E6;
 $background-color: #FAFAFA;
 $main-beige: #FCF9F2;
 $dark-grey: #707070;
+$main-turquoise: #233A26;
